+++ tests/tests_asian_pricing.cpp
@@ -7,8 +7,8 @@
 #include <memory>
 /*  -- TEST BOILERPLATE --  */
 using namespace simdmonte;
-const static int TEST_SIMS = 1e7;
-const static int TEST_STEPS = 252; // TODO: Move this into test params to test different steps
+const static long TEST_SIMS = 1e7;
+const static int TEST_STEPS = 252;
 
 struct AsianTestCase {
   std::string name;
